backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# HTTP clients for microservice communication
httpx==0.25.2
aiohttp==3.9.1

# AI/ML Libraries
torch==2.1.1
transformers==4.36.2
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
sentence-transformers==2.2.2

# NLP Processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1

# Vector Databases
chromadb==0.4.18
pinecone-client==2.2.4
weaviate-client==3.25.3

# Speech to Text
openai==1.3.7
azure-cognitiveservices-speech==1.34.0

# Graph Database
neo4j==5.14.1
py2neo==2021.2.4

# Clustering and Analysis
plotly==5.17.0
matplotlib==3.8.2
seaborn==0.13.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring and Observability
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Email and Notifications
emails==0.6.0
celery==5.3.4
flower==2.0.1

# Utilities
python-dotenv==1.0.0
pyjwt==2.8.0
requests==2.31.0
click==8.1.7
pytz==2023.3

# Image processing (for future avatar features)
Pillow==10.1.0

# WebSocket support
websockets==12.0 